requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
selenium==4.15.2
pandas==2.0.3
numpy==1.24.3
google-cloud-bigquery==3.13.0
python-dotenv==1.0.0
webdriver-manager==4.0.1
fake-useragent==1.4.0
schedule==1.2.1 